refactor(character, board): include state and service in related module
