+++ src/app/board/board.model.ts
@@ -1,0 +1,3 @@
+export interface Board {
+  tiles: string[];
+}
