+++ src/app/board/components/board/board.component.scss
