+++ src/app/board/state/board.reducer.ts
